pfm/pyproject.toml
[project]
name = "pfm"
version = "0.1.0"
description = "Personal finance management app — categorization engine and API"
requires-python = ">=3.12"
dependencies = [
    # API framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Cache & message broker
    "redis[hiredis]>=5.2.0",

    # Background tasks
    "celery[redis]>=5.4.0",

    # Auth
    "PyJWT[crypto]>=2.9.0",

    # Logging
    "structlog>=25.0.0",
    "asgi-correlation-id>=4.3.4",

    # LLM
    "anthropic>=0.42.0",

    # Observability
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-instrumentation-redis>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",

    # Analytics
    "posthog>=3.7.0",

    # Plaid
    "plaid-python>=27.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
pipeline = [
    "duckdb>=1.1.0",
    "pandas>=2.2.0",
    "pyarrow>=18.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "pipeline"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
